\name{kfsang-package}
\alias{kfsang-package}
\alias{kfsang}
\docType{package}
\title{
\packageTitle{kfsang}
}
\description{
\packageDescription{kfsang}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kfsang}
\packageIndices{kfsang}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{kfsang}

Maintainer: \packageMaintainer{kfsang}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
